feat(frontend/2FA_panel) : Done with 2FA enabling flow
