+++ frontend/src/components/2FA_panel.html
@@ -24,6 +24,6 @@
         </div>
       </form>
     </div>
-  <button id="confirm-2fa-button" class="button button-md">Submit</button>
+  <button id="submit-2fa-button" class="button button-md">Submit</button>
   </div>
 </div>
