+++ frontend/src/js/main_menu_panel.js
@@ -3,7 +3,7 @@ import { loadComponent } from "./ui_utils/ui_utils.js";
 import { initHotbar } from "./ui_utils/hotbar_utils.js";
 import { initAddFriendButton, loadToFriendsContainer, initCloseSearchFriendButton } from "./friends_content.js";
 import { initBackButton, initRandomColorButton } from "./ui_utils/button_utils.js"
-import { send_otp_2FA } from "./network_utils/token_utils.js";
+import { send_otp_2FA } from "./network_utils/2FA_utils.js";
 import { load2FAPanel } from "./2FA_panel.js";
 // document.addEventListener("DOMContentLoaded", () => {
 //   loadMainMenuPanel()
