+++ frontend/src/js/network_utils/token_utils.js
@@ -14,18 +14,3 @@ export async function getIdToken(loginInfo) {
     alert('Token Creation Error');
   }
 }
-
-export async function send_otp_2FA() {
-  try {
-    const response = await postRequest('/two_factor_auth/send_otp_2FA/')
-
-    if (response.success) {
-      alert('2FA code sent to ur email');
-    } else {
-      alert('An error occurred. Please try again.');
-    }
-  } catch (error) {
-    console.error('Error:', error);
-    alert('An error occurred. Please try again.');
-  }
-}
